Move file types to their own module
