+++ pysyte/types/paths.py
@@ -746,60 +746,6 @@ class DirectPath(DotPath, PathAssertions):
         return str(self) == "/"
 
 
-@dataclass
-class FileType:
-    """A file extension and associated type
-
-    >>> from pym.ast import AST
-    >>> FileType(AST, "py")
-    """
-    type_: type
-    ext: str
-
-    def extend(self, file_stem: StrPath) -> FilePath:
-        """Returns file stem with missing extension added."""
-        return file_stem.add_missing_ext(self.ext)
-
-    def typed(self, file_stem: StrPath) -> Any: # actually returns instance of self.type_, but mypy is hard
-        """Adds extension to stem and attempts to cast it to the associated type."""
-        file = self.extend(file_stem)
-        if file:
-            return self.type_(file)
-        return None
-
-
-@dataclass
-class FileTypes:
-    """A collection of FileTypes
-
-    Assuming "fred.yml", or "fred.yaml" exists, then
-    >>> class YamlConfig:
-    >>>     pass
-    >>> configs = FileTypes((YamlConfig, "yaml"), (YamlConfig, "yml"))
-    >>> fred = configs.typed("fred")
-    >>> assert isinstance(fred, YamlConfig)
-    """
-    file_types: List[FileType]
-
-    def __init__(self, file_types: List[FileType]):
-        """Convert all elements in file_types to FileType instances."""
-        self.file_types = [_ if isinstance(_, FileType) else FileType(*_) for _ in file_types]
-
-    def extend(self, file_stem: FilePath) -> list[FilePath]
-        """The first real file matching the stem with one of our extensions"""
-        for file_type in self.file_types:
-            file = file_type.extend(file_stem)
-            if file:
-                yield file
-
-    def typed(self, file_stem: FilePath) -> Any: # actually yields instances of self.file_types, but mypy is still hard 
-        """An iterator of all possible extensions of that stem which are real files"""
-        for file_type in self.file_types:
-            typed = file_type.typed(file_stem)
-            if typed:
-                yield typed
-
-
 def ignore_fnmatches(ignores):
     def ignored(a_path):
         if not ignores:
